added corner icon
