+++ mysite/matches/models.py
@@ -2,21 +2,10 @@ from django.db import models
 from django.utils import timezone
 from django_enumfield import enum
 from django.contrib.auth.models import User
+
 from products.models import Product
+from .model_enums import OfferStatus
 
-
-class OfferStatus(enum.Enum):
-    PENDING = 0  # pending review
-    REJECTED = 1
-    ACCEPTED = 2
-
-    __default__ = PENDING
-
-    # InvalidStatusOperationError exception will be raised if we attempt an invalid transition
-    __transitions__ = {
-        REJECTED: (PENDING,),  # Can go from PENDING to REJECTED
-        ACCEPTED: (PENDING,)  # Can go from PENDING to ACCEPTED
-    }
 
 
 class Offer(models.Model):
