+++ mysite/matches/templates/matches/make_offer.html
@@ -5,7 +5,7 @@
             {% csrf_token %}
             <div class="form-group" style="text-align: center">
                 <h2>Select which items you want to offer</h2>
-                <button class="btn btn-outline-info" type="submit">Submit</button>
+                <button class="btn btn-outline-info" type="submit">Send Offer</button>
             </div>
             {% include 'matches/_make_offer_feed_component.html' %}
         </form>
