+++ mysite/products/templates/products/_base_feed_component.html
@@ -8,7 +8,7 @@
 </div>
 <script>
     $(document).ready(function() {
-        let images = $('.grid-item');
+        let images = $('.grid-item, .offer-status-icon');
         let i
         for (i=0; i<images.length; i++){
             images[i].style.visibility = 'visible'
